+++ src/common/components/index.js
@@ -1,2 +1,3 @@
 export { default as ApprovedBy } from './ApprovedBy';
+export { default as DuplicateInvoiceList } from './DuplicateInvoiceList';
 export * from './Status';